+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/bot/usecase/CreateEmbedUseCase.kt
@@ -16,6 +16,7 @@ import io.github.aakira.napier.Napier
 import io.github.sophon.core.architecture.Result
 import io.github.sophon.core.util.rollChance
 import io.github.sophon.discord.EMBED_BUTTON_DURATION_INF
+import io.github.sophon.discord.RNG_DONATION_PCT_COMMAND
 import io.github.sophon.discord.URL_KOFI
 import io.github.sophon.discord.feat.core.domain.DiscordButtonBuilder
 import io.github.sophon.discord.feat.core.domain.model.BotError
@@ -83,7 +84,7 @@ internal class CreateEmbedUseCase(
                 }
             }
 
-            if (rollChance(successPercentage = 5)) {
+            if (rollChance(successPercentage = RNG_DONATION_PCT_COMMAND)) {
                 message.channel.createMessage {
                     content = donationMessage()
                 }
@@ -132,9 +133,9 @@ internal class CreateEmbedUseCase(
                 }
             }
 
-            if (rollChance(successPercentage = 1)) {
+            if (rollChance(successPercentage = RNG_DONATION_PCT_COMMAND)) {
                 interaction.channel.createMessage {
-                    content = "Consider donating (`/donate` or `/tip`): **<${URL_KOFI}>**"
+                    content = donationMessage()
                 }
             }